Add blank rows and columns detection

+++ segmentation/threshold.m
@@ -2,7 +2,7 @@ function [ imout ] = threshold( im )
 %THRESHOLD Summary of this function goes here
 %   Detailed explanation goes here
 
-imout = im < 0.5;
+imout = double(im < 0.5);
 
 end
